refactor: extract ComputeDraftCard from ConversationsPanel (#672)
Pulls the propose_compute review card (#245) — the richest of the draft cards:
language pill, rationale, safety flags, inline code editor, Run / Edit / Insert
/ Discard, and the run output (text / json / table / image / html) — into its
own ComputeDraftCard.svelte. ConversationsPanel 2162 → 1824.

This also clears the temporary chrome-CSS duplication the DraftCard PR flagged:
with the compute card gone, the panel no longer renders any .draft-card /
.draft-note / .draft-actions / .draft-btn, so all of that (plus the whole
.compute-* block) is deleted from ConversationsPanel. Only the simple cards'
body styles (.draft-paths, source/property/claims lists) remain — they render
as DraftCard children, in the panel's scope.

Design: the card owns its ephemeral view state — edit buffer, edit mode, and
the risky-Python "click Run twice" acknowledgement — while the panel keeps the
run state (running / result / insertedAt, in the conversation store) and the
three store actions (Run / Insert / Discard) + open-inserted, passed as
callbacks. The Trust boundary is unchanged: the card only proposes.

Tests (9): components/ComputeDraftCard.test.ts pins the render, the risky two-
click confirmation, edited-code-folds-into-Run, Insert/Discard routing, the
running/disabled state, error + table output, and the filed-cell link.

Verified: pnpm lint (0 errors), full suite (2980 passed, +9), electron-forge
build + boot smoke.
